Enable UnsafeTypeCastWarningLevel=Error; fix death handling; wire DamageCauser
- Build.cs: implicit int<->float narrowing (C4244/C4838) now fails the build
  for the game module; engine default suppresses it. Fallout fixed: explicit
  casts at the Enhanced Input call sites, anim GroundSpeed, and the int
  attribute math; EnhancedInput headers (engine-owned narrowing) wrapped in
  PRAGMA_DISABLE/RESTORE_UNSAFE_TYPECAST_WARNINGS.
- AttributeComponent: CharacterDied now fires once, on the alive->dead
  transition (corpse hits used to re-ragdoll and re-fire it every hit).
- BP_MiliBot/BP_AoE: connect Self -> DamageCauser on Apply Damage, so block
  verdicts get the attacker position (AoE was blockable from behind).
- Combat gym: BP_GymRespawner reloads the level on player death (no more
  soft-lock); placed in L_CombatGym.

+++ Source/RhyaTowerOfWishes/Private/Characters/Character_AnimInstance.cpp
@@ -22,7 +22,7 @@ void UCharacter_AnimInstance::NativeUpdateAnimation(float DeltaTime)
 
 	if (characterMovementComponent)
 	{
-		GroundSpeed = UKismetMathLibrary::VSizeXY(characterMovementComponent->Velocity);
+		GroundSpeed = (float)UKismetMathLibrary::VSizeXY(characterMovementComponent->Velocity);
 		IsFalling = characterMovementComponent->IsFalling();
 		characterState = character->GetCharacterState();
 	}
